storyteller-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
requests
python-dotenv
google-cloud-storage
google-cloud-texttospeech
